+++ main.go
@@ -2,22 +2,26 @@ package main
 
 import (
 	"fmt"
-	"github.com/manifoldco/promptui"
-	"github.com/manifoldco/promptui/list"
 	"log"
 	"os"
 	"path/filepath"
 	"sort"
 	"strings"
+
+	"github.com/manifoldco/promptui"
+	"github.com/manifoldco/promptui/list"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/client-go/tools/clientcmd/api"
 )
 
 const (
-	NoticeColor = "\033[0;38m%s\u001B[0m"
-	PromptColor = "\033[1;38m%s\u001B[0m"
-	CyanColor   = "\033[0;36m%s\033[0m"
+	NoticeColor  = "\033[0;38m%s\u001B[0m"
+	PromptColor  = "\033[1;38m%s\u001B[0m"
+	CyanColor    = "\033[0;36m%s\033[0m"
+	MagentaColor = "\033[0;35m%s\033[0m"
 )
 
-var version string = "v0.0.2"
+var version string = "v0.0.3"
 
 func newPromptUISearcher(items []string) list.Searcher {
 	return func(searchInput string, itemIndex int) bool {
@@ -26,14 +30,61 @@ func newPromptUISearcher(items []string) list.Searcher {
 }
 
 func main() {
-	if len(os.Args) > 1 && os.Args[1] == "version" {
-		fmt.Println("kxd version", version)
-		os.Exit(0)
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		log.Fatalf("Error getting user home directory: %v\n", err)
 	}
-	home := os.Getenv("HOME")
-	configFileLocation := fmt.Sprintf("%s/.kube", home)
-	configs := getConfigs(configFileLocation)
-	err := touchFile(fmt.Sprintf("%s/.kxd", home))
+
+	if len(os.Args) > 1 {
+		arg := strings.ToLower(os.Args[1])
+		switch arg {
+		case "-v", "--v", "version":
+			fmt.Println("kxd version:", version)
+		case "-c", "--c", "context":
+			err := runContextSwitcher(homeDir)
+			if err != nil {
+				log.Fatal(err)
+			}
+		case "-h", "--h", "help":
+			err := displayHelp()
+			if err != nil {
+				log.Fatal(err)
+			}
+		default:
+			err := runConfigSwitcher(homeDir)
+			if err != nil {
+				log.Fatal(err)
+			}
+		}
+	} else {
+		err := runConfigSwitcher(homeDir)
+		if err != nil {
+			log.Fatal(err)
+		}
+	}
+}
+
+func displayHelp() error {
+	var helpMessage strings.Builder
+	options := map[string]string{
+		"no operation": "Switch configs.",
+		"       -c   ": "Switch contexts.",
+		"       -h   ": "Help. Displays this message.",
+		"       -v   ": "Displays version.",
+	}
+	helpMessage.WriteString("Usage: kxd [OPERATION]\n")
+	for option, description := range options {
+		helpMessage.WriteString(fmt.Sprintf("  %s: %s\n", option, description))
+	}
+	fmt.Println(helpMessage.String())
+	return nil
+}
+
+func runConfigSwitcher(homeDir string) error {
+	configFileLocation := fmt.Sprintf("%s/.kube", homeDir)
+	configs := getConfigs(configFileLocation, homeDir)
+
+	err := touchFile(fmt.Sprintf("%s/.kxd", homeDir))
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -54,20 +105,88 @@ func main() {
 		StartInSearchMode: true,
 		Stdout:            &bellSkipper{},
 	}
-	_, result, err := prompt.Run()
 
+	_, result, err := prompt.Run()
 	if err != nil {
 		checkError(err)
-		return
 	}
+
 	fmt.Printf(PromptColor, "Choose a config")
 	fmt.Printf(NoticeColor, "? ")
 	fmt.Printf(CyanColor, result)
 	fmt.Println("")
+
 	if result == "default" {
 		result = "config"
 	}
-	writeFile(result, home)
+	writeFile(result, homeDir)
+
+	return nil
+}
+
+func runContextSwitcher(homeDir string) error {
+	kubeconfigPath := getenv("KUBECONFIG", filepath.Join(homeDir, ".kube/config"))
+	config, err := initializeKubeconfig(kubeconfigPath)
+	if err != nil {
+		log.Fatalf("Error initializing kubeconfig: %v\n", err)
+	}
+
+	contexts := listContexts(kubeconfigPath)
+
+	fmt.Printf(NoticeColor, "Kubeconfig Context Switcher\n")
+	prompt := promptui.Select{
+		Label:        fmt.Sprintf(PromptColor, "Choose a context"),
+		Items:        contexts,
+		HideHelp:     true,
+		HideSelected: true,
+		Templates: &promptui.SelectTemplates{
+			Label:    "{{ . }}?",
+			Active:   fmt.Sprintf("%s {{ . | magenta }}", promptui.IconSelect),
+			Inactive: "  {{.}}",
+			Selected: "  {{ . | magenta }}",
+		},
+		Searcher:          newPromptUISearcher(contexts),
+		StartInSearchMode: true,
+		Stdout:            &bellSkipper{},
+	}
+
+	_, result, err := prompt.Run()
+	if err != nil {
+		checkError(err)
+	}
+
+	fmt.Printf(PromptColor, "Choose a context")
+	fmt.Printf(NoticeColor, "? ")
+	fmt.Printf(MagentaColor, result)
+	fmt.Println("")
+
+	err = switchContext(config, result, kubeconfigPath)
+	if err != nil {
+		log.Fatalf("Error switching context: %v\n", err)
+	}
+	return nil
+}
+
+func initializeKubeconfig(kubeconfigPath string) (*api.Config, error) {
+	return clientcmd.LoadFromFile(kubeconfigPath)
+}
+
+func listContexts(kubeconfigPath string) []string {
+	config, err := initializeKubeconfig(kubeconfigPath)
+	if err != nil {
+		log.Fatalf("Error initializing kubeconfig: %v\n", err)
+	}
+
+	var contexts []string
+	for contextName := range config.Contexts {
+		contexts = append(contexts, contextName)
+	}
+	return contexts
+}
+
+func switchContext(config *api.Config, contextName string, kubeconfigPath string) error {
+	config.CurrentContext = contextName
+	return clientcmd.WriteToFile(*config, kubeconfigPath)
 }
 
 func touchFile(name string) error {
@@ -89,26 +208,20 @@ func writeFile(config, loc string) {
 	}
 }
 
-func getConfigs(configFileLocation string) []string {
+func getConfigs(configFileLocation string, homeDir string) []string {
 	var files []string
 	fileExt := getenv("KXD_MATCHER", ".conf")
 	err := filepath.Walk(configFileLocation, func(path string, f os.FileInfo, _ error) error {
-		if !f.IsDir() {
-			if strings.Contains(f.Name(), fileExt) {
-				files = append(files, f.Name())
-			}
+		if !f.IsDir() && strings.Contains(f.Name(), fileExt) {
+			files = append(files, f.Name())
 		}
 		return nil
 	})
 	if err != nil {
 		log.Fatal(err)
 	}
-	if len(files) < 1 {
-		fmt.Println("No matching config files found.")
-		os.Exit(1)
-	}
-	_, err = os.Stat(fmt.Sprintf("%s/.kube/config", os.Getenv("HOME")))
-	if err == nil {
+
+	if _, err := os.Stat(fmt.Sprintf("%s/.kube/config", homeDir)); err == nil {
 		files = append(files, "default")
 	}
 	files = append(files, "unset")